requirements.txt
flask==2
psycopg2-binary
Flask-SQLAlchemy
requests
APScheduler==3.3.1
pytest